+++ R/brmsfit-class.R
@@ -26,6 +26,8 @@
 #' @slot threads An object of class `brmsthreads` created by
 #'   \code{\link{threading}}.
 #' @slot opencl An object of class `brmsopencl` created by \code{\link{opencl}}.
+#' @slot stan_args Named list of additional control arguments that were passed
+#'   to the Stan backend directly.
 #' @slot fit An object of class \code{\link[rstan:stanfit-class]{stanfit}}
 #'   among others containing the posterior draws.
 #' @slot criteria An empty \code{list} for adding model fit criteria
@@ -56,7 +58,7 @@ brmsfit <- function(formula = NULL, data = data.frame(), prior = empty_prior(),
                     ranef = empty_ranef(), save_pars = NULL,
                     algorithm = "sampling", backend = "rstan",
                     threads = threading(), opencl = opencl(),
-                    fit = NULL, criteria = list(),
+                    stan_args = list(), fit = NULL, criteria = list(),
                     file = NULL, family = NULL, autocor = NULL,
                     cov_ranef = NULL, stan_funs = NULL, data.name = "") {
   version <- list(
@@ -71,8 +73,8 @@ brmsfit <- function(formula = NULL, data = data.frame(), prior = empty_prior(),
   }
   x <- nlist(
     formula, data, prior, data2, stanvars, model, ranef,
-    save_pars, algorithm, backend, threads, opencl, fit, criteria, file,
-    version, family, autocor, cov_ranef, stan_funs, data.name
+    save_pars, algorithm, backend, threads, opencl, stan_args, fit, criteria,
+    file, version, family, autocor, cov_ranef, stan_funs, data.name
   )
   class(x) <- "brmsfit"
   x